优化

+++ feedback-starter/src/main/java/com/eighteen/common/feedback/service/FeedbackService.java
@@ -17,7 +17,7 @@ public interface FeedbackService {
 
     void clean(JobType type);
 
-    void clearCache(Integer offset);
+    void clearCache(Long offset);
 
     void stat(JobType type);
